+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Support `DoubleEndedIterator` for `IntervalSet.iter()`.
+
 ## [0.1.1] - 2023-04-23
 
 - Documentation updates.